fix(keys): renamed file to serde_public_key.rs
