+++ crates/iota-keys/src/lib.rs
@@ -7,4 +7,4 @@ pub mod keypair_file;
 pub mod keystore;
 pub mod random_names;
 mod serde_iota_keypair;
-mod serde_publickey;
+mod serde_public_key;
